* server/asobj/SoundFfmpeg.cpp: Fix signed/unsigned compiler warning

+++ server/asobj/SoundFfmpeg.cpp
@@ -151,7 +151,7 @@ SoundFfmpeg::setupDecoder(SoundFfmpeg* so)
 	// Find the first audio stream
 	so->audioIndex = -1;
 
-	for (unsigned int i = 0; i < so->formatCtx->nb_streams; i++)
+	for (unsigned int i = 0; i < (unsigned)so->formatCtx->nb_streams; i++)
 	{
 		AVCodecContext* enc = so->formatCtx->streams[i]->codec;
